requirements.txt
# LINE Messaging API
line-bot-sdk==3.5.0

# Google APIs
gspread==5.12.0
google-auth==2.25.2
google-api-python-client==2.110.0
google-cloud-storage==2.14.0

# 圖片處理
Pillow==10.1.0

# 時區處理
pytz==2023.3

# Web 框架
Flask==3.0.0

# Production WSGI Server
gunicorn==21.2.0

# 環境變數管理
python-dotenv==1.0.0

# 測試
pytest==7.4.3
pytest-cov==4.1.0
requests-mock==1.11.0
